Energies

Hydrogen.non_relativistic_energy — Function

non_relativistic_energy(n[, Z=1])

Returns the non-relativistic energy of a hydrogen orbital with principal quantum number $n$ (and an effective nuclear charge $Z$) according to

\[\begin{equation} \tag{BS2.11} E_n = -\frac{Z^2}{2n^2} \end{equation}\]

non_relativistic_energy(o)

Return the non-relativistic energy of the orbital o.

Examples

julia> non_relativistic_energy(o"1s")
-0.5

julia> non_relativistic_energy(o"2p")
-0.125

julia> non_relativistic_energy(ro"2p")
-0.125

julia> non_relativistic_energy(ro"2p-")
-0.125

Hydrogen.relativistic_energy — Function

relativistic_energy(n, κ[, Z=1])

Returns the relativistic energy of a hydrogen orbital with principal quantum number $n$ and "angular momentum" quantum number $κ$ (and an effective nuclear charge $Z$) according to

\[\begin{equation} \tag{BS14.29*} E_{nκ} = c^2\left[ -1+ \frac{1}{\sqrt{1 + \left( \frac{Zα}{n - \abs{κ} + \sqrt{κ^2-Z^2α^2}} \right)^2}} \right] \end{equation}\]

where the energy has been shifted by $-c^2$ compared to $\mathrm{(BS14.29)}$.

relativistic_energy(o::RelativisticOrbital)

Return the relativistic energy for the orbital o.

Example

julia> relativistic_energy(ro"1s")
-0.5000066565961734

julia> relativistic_energy(ro"2p-")
-0.12500208019062972

julia> relativistic_energy(ro"2p")
-0.1250004160284539

Hydrogen.orbital_energy — Function

orbital_energy(o)

Return the energy of orbital o, automatically choosing between non_relativistic_energy and relativistic_energy, depending on the type of o.

Examples

julia> orbital_energy(o"1s")
-0.5

julia> orbital_energy(ro"1s")
-0.5000066565961734

julia> orbital_energy(o"2p")
-0.125

julia> orbital_energy(ro"2p")
-0.1250004160284539

Hydrogen.atomic_hamiltonian — Function

atomic_hamiltonian(orbitals[; Efun=orbital_energy, neutral_zero=true])

Return the diagonal matrix whose elements corresponds to the energies of each of the orbitals, by default orbital_energy is used to determine the energy, which dispatches on the orbital type to return the appropriate energy (non-relativistic or relativistic), and all energies are shifted such that 1s has zero energy, if neutral_zero==true.

Examples

julia> atomic_hamiltonian(os"1[s] 2[s-p]")
3×3 LinearAlgebra.Diagonal{Float64,Array{Float64,1}}:
 0.0   ⋅      ⋅
  ⋅   0.375   ⋅
  ⋅    ⋅     0.375

julia> atomic_hamiltonian(ros"1[s] 2[s-p]")
4×4 LinearAlgebra.Diagonal{Float64,Array{Float64,1}}:
 0.0   ⋅         ⋅         ⋅
  ⋅   0.375005   ⋅         ⋅
  ⋅    ⋅        0.375005   ⋅
  ⋅    ⋅         ⋅        0.375006